Set up cosmos validation on backend
