+++ client/scripts/controllers/chain/cosmos/account.ts
@@ -4,7 +4,6 @@ import { first } from 'rxjs/operators';
 
 import { IApp } from 'state';
 import { CosmosToken } from 'adapters/chain/cosmos/types';
-import { keyToMsgSend, VALIDATION_CHAIN_DATA } from 'adapters/chain/cosmos/keys';
 import CosmosChain from 'controllers/chain/cosmos/chain';
 import { Account, IAccountsModule } from 'models';
 import { AccountsStore } from 'stores';
