+++ shared/adapters/chain/cosmos/keys.ts
@@ -1,13 +1,8 @@
 // Cosmos cannot sign arbitrary blobs, but they can sign transactions. So, as a hack around that,
 // we insert our account registration token into a proposal message, and then verify against the
 // generated signature. But first we need the message to insert.
-export const VALIDATION_CHAIN_DATA = {
-  sequence: '0',
-  accountNumber: '0',
-  chainId: 'straightedge-2',
-};
 
-export const keyToMsgSend = async (address: string, token: string) => {
+export const validationTokenToSignDoc = async (address: string, token: string) => {
   const CosmosApi = await import('@lunie/cosmos-api');
   const cosmos = new CosmosApi.default('');
   const jsonTx = (cosmos as any).MsgSubmitProposal(address, {
@@ -21,11 +16,16 @@ export const keyToMsgSend = async (address: string, token: string) => {
     gasPrices: [{ denom: 'stake', amount: '0' }],
     memo: ''
   }, jsonTx);
-  return stdTx;
-};
+  delete stdTx.fee;
+  delete stdTx.memo;
 
-export const keyToSignMsg = async (address: string, token: string) => {
-  const stdTx = await keyToMsgSend(address, token);
-  // TODO: swap out this chain ID either for a constant or something that can be retrieved on the server
-  return (await import('@lunie/cosmos-api')).createSignMessage(stdTx, VALIDATION_CHAIN_DATA);
+  const signDoc = {
+    chain_id: 'straightedge-2',
+    account_number: '0',
+    sequence: '0',
+    fee: { gas: '100000', amount: [{ denom: 'astr', amount: '2500000000000000' }] },
+    memo: '',
+    msgs: stdTx.msg,
+  };
+  return signDoc;
 };
